+++ CHANGELOG.md
@@ -1,7 +1,10 @@
+## 0.0.3
+- Photo capture fix
+
+## 0.0.2
+- ReadMe description changed
+
 ## 0.0.1
 - Add CameraWidget and CameraController on Android.
 - Add camerax example.
 - Face detection
-
-## 0.0.2
-- ReadMe description changed